requirements.txt
python-telegram-bot[job-queue]>=20.7
aiosqlite>=0.20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
